requirements.txt
aws-cdk-lib==2.40.0
constructs>=10.0.0,<11.0.0
boto3==1.24.67
